hide `inner` from interface.
