+++ azure-spring-cloud-maven-plugin/src/test/java/com/microsoft/azure/maven/springcloud/TestHelper.java
@@ -15,7 +15,7 @@ import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.Ser
 import com.microsoft.azure.management.resources.Subscription;
 import com.microsoft.azure.management.resources.Subscriptions;
 import com.microsoft.azure.maven.telemetry.AppInsightHelper;
-import com.microsoft.azure.toolkit.lib.springcloud.model.SpringCloudClusterEntity;
+import com.microsoft.azure.toolkit.lib.springcloud.SpringCloudClusterEntity;
 import com.microsoft.rest.RestException;
 
 import org.apache.maven.model.Model;
