+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/SpringCloudClusterEntity.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package com.microsoft.azure.toolkit.lib.springcloud.model;
+package com.microsoft.azure.toolkit.lib.springcloud;
 
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.ServiceResourceInner;
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.SkuInner;
@@ -35,6 +35,7 @@ import org.apache.commons.lang3.ArrayUtils;
 public class SpringCloudClusterEntity implements IAzureEntity {
     private final String resourceGroup;
     private final String name;
+    @Getter(AccessLevel.PACKAGE)
     private ServiceResourceInner inner;
 
     private SpringCloudClusterEntity(ServiceResourceInner resource) {
