+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/service/SpringCloudClusterManager.java
@@ -11,8 +11,8 @@ import com.microsoft.azure.management.appplatform.v2020_07_01.ServiceResource;
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.AppPlatformManager;
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.AppResourceInner;
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.ServiceResourceInner;
-import com.microsoft.azure.toolkit.lib.springcloud.model.SpringCloudAppEntity;
-import com.microsoft.azure.toolkit.lib.springcloud.model.SpringCloudClusterEntity;
+import com.microsoft.azure.toolkit.lib.springcloud.SpringCloudAppEntity;
+import com.microsoft.azure.toolkit.lib.springcloud.SpringCloudClusterEntity;
 
 import java.util.List;
 import java.util.Objects;
